requirements.txt
Django>=5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
gunicorn>=21.2
whitenoise>=6.6
